# docs/04-api/11-OpenAPI-Core-v0.4.yaml
openapi: 3.0.3
info:
  title: Pet Platform V1.0 Core API
  version: 0.4.0
  description: '宠物平台 V1.0 核心交易 OpenAPI 草案。 完整业务语义以产品 SSOT、技术基线与 HTTP Controller Contract
    为准。

    '
servers:
- url: /api/v1
...
paths:
...
  /c/auth/attempts:
    post:
      operationId: cAuthCreateAttempt
      tags:
      - AUTH-Miniapp
      summary: cAuthCreateAttempt
      description: 'AUTH-001 accepted draft-v2 mapping; sync candidate, not implemented. '
      security: []
      parameters:
      - $ref: '#/components/parameters/RequestId'
      responses:
        '400':
          description: Invalid arguments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '401':
          description: Credential invalid, expired or revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '403':
          description: Action/scope denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '404':
          description: Not found or anti-enumeration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '409':
          description: Idempotency/version/state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '503':
          description: Authoritative dependency unavailable; do not invent success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '201':
          description: Contract result, not deployed implementation. Initial attempt secret is never replayed by UUID
            alone.
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CAuthCreateAttemptResponse'
      x-auth-mode: anonymous-bootstrap
      x-auth-audience: MINIAPP
      x-contract-status: SYNC_CANDIDATE_NOT_IMPLEMENTED
      x-approval-source: 2963e3918bbd62afdfdeb55c60bb2e2221c67ca1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthMiniAttemptRequest'
      x-secret-replay: NEVER_BY_REQUEST_ID_ALONE
...
  /c/auth/session:
    get:
      operationId: cAuthGetSession
      tags:
      - AUTH-Miniapp
      summary: cAuthGetSession
      description: 'AUTH-001 accepted draft-v2 mapping; sync candidate, not implemented. '
      security:
      - bearerAuth: []
      parameters: []
      responses:
        '400':
          description: Invalid arguments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '401':
          description: Credential invalid, expired or revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '403':
          description: Action/scope denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '404':
          description: Not found or anti-enumeration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '409':
          description: Idempotency/version/state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '503':
          description: Authoritative dependency unavailable; do not invent success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '200':
          description: 'Contract result, not deployed implementation. '
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CAuthGetSessionResponse'
      x-auth-mode: bearer
      x-auth-audience: MINIAPP
      x-contract-status: SYNC_CANDIDATE_NOT_IMPLEMENTED
      x-approval-source: 2963e3918bbd62afdfdeb55c60bb2e2221c67ca1
...
  /c/auth/logout:
    post:
      operationId: cAuthLogout
      tags:
      - AUTH-Miniapp
      summary: cAuthLogout
      description: 'AUTH-001 accepted draft-v2 mapping; sync candidate, not implemented. '
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/RequestId'
      responses:
        '400':
          description: Invalid arguments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '401':
          description: Credential invalid, expired or revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '403':
          description: Action/scope denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '404':
          description: Not found or anti-enumeration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '409':
          description: Idempotency/version/state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '503':
          description: Authoritative dependency unavailable; do not invent success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '200':
          description: 'Contract result, not deployed implementation. '
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CAuthLogoutResponse'
      x-auth-mode: bearer
      x-auth-audience: MINIAPP
      x-contract-status: SYNC_CANDIDATE_NOT_IMPLEMENTED
      x-approval-source: 2963e3918bbd62afdfdeb55c60bb2e2221c67ca1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthEmptyRequest'
  /c/auth/attempts/{attemptId}/result:
    get:
      operationId: cAuthGetAttemptResult
      tags:
      - AUTH-Miniapp
      summary: cAuthGetAttemptResult
      description: 'AUTH-001 accepted draft-v2 mapping; sync candidate, not implemented. '
      security:
      - authAttempt: []
      parameters:
      - name: attemptId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/PublicId'
      - $ref: '#/components/parameters/AuthOriginalRequestId'
      responses:
        '400':
          description: Invalid arguments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '401':
          description: Credential invalid, expired or revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '403':
          description: Action/scope denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '404':
          description: Not found or anti-enumeration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '409':
          description: Idempotency/version/state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '503':
          description: Authoritative dependency unavailable; do not invent success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '200':
          description: 'Contract result, not deployed implementation. '
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CAuthGetAttemptResultResponse'
      x-auth-mode: attempt
      x-auth-audience: MINIAPP
      x-contract-status: SYNC_CANDIDATE_NOT_IMPLEMENTED
      x-approval-source: 2963e3918bbd62afdfdeb55c60bb2e2221c67ca1
...
  /admin/auth/attempts:
    post:
      operationId: adminAuthCreateAttempt
      tags:
      - AUTH-Admin
      summary: adminAuthCreateAttempt
      description: "Accepted AUTH Web component candidate; disabled by default. Production host, secrets and migration enablement remain gated. "
      security: []
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/AuthOrigin'
      responses:
        '400':
          description: Invalid arguments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '401':
          description: Credential invalid, expired or revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '403':
          description: Action/scope denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '404':
          description: Not found or anti-enumeration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '409':
          description: Idempotency/version/state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '503':
          description: Authoritative dependency unavailable; do not invent success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '201':
          description: Contract result, not deployed implementation. Initial attempt secret is never replayed by UUID
            alone.
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
            Set-Cookie:
              schema:
                type: string
              description: __Host-pet-admin-attempt request binding; Secure; HttpOnly; SameSite=Strict; Path=/; never
                log cookie value.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAuthCreateAttemptResponse'
      x-auth-mode: anonymous-bootstrap
      x-auth-audience: ADMIN_WEB
      x-contract-status: CONTRACT_SYNCED_IMPLEMENTATION_CANDIDATE
      x-approval-source: 2963e3918bbd62afdfdeb55c60bb2e2221c67ca1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthEmptyRequest'
      x-secret-replay: NEVER_BY_REQUEST_ID_ALONE
  /admin/auth/login:
    post:
      operationId: adminAuthLogin
      tags:
      - AUTH-Admin
      summary: adminAuthLogin
      description: "Accepted AUTH Web component candidate; disabled by default. Production host, secrets and migration enablement remain gated. All roles complete ordinary password authentication; no additional factor. New session/generation atomically invalidates older login. B1 secret recovery expires at pre-commit database anchor plus 60 seconds, possibly less than 60 seconds after commit. Cache/key loss returns 503; expired or revoked proof returns 401; never reissue or extend the session."
      security:
      - authAttempt: []
        adminBinding: []
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/AuthOrigin'
      responses:
        '400':
          description: Invalid arguments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '401':
          description: Credential invalid, expired or revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '403':
          description: Action/scope denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '404':
          description: Not found or anti-enumeration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '409':
          description: Idempotency/version/state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '503':
          description: Authoritative dependency unavailable; do not invent success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '200':
          description: 'Contract result, not deployed implementation. '
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAuthLoginResponse'
      x-auth-mode: admin-attempt
      x-auth-audience: ADMIN_WEB
      x-contract-status: CONTRACT_SYNCED_IMPLEMENTATION_CANDIDATE
      x-approval-source: 2963e3918bbd62afdfdeb55c60bb2e2221c67ca1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthWebLoginRequest'
...
  /admin/auth/session:
    get:
      operationId: adminAuthGetSession
      tags:
      - AUTH-Admin
      summary: adminAuthGetSession
      description: "Accepted AUTH Web component candidate; disabled by default. Production host, secrets and migration enablement remain gated. "
      security:
      - bearerAuth: []
      parameters: []
      responses:
        '400':
          description: Invalid arguments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '401':
          description: Credential invalid, expired or revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '403':
          description: Action/scope denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '404':
          description: Not found or anti-enumeration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '409':
          description: Idempotency/version/state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '503':
          description: Authoritative dependency unavailable; do not invent success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorEnvelope'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '200':
          description: 'Contract result, not deployed implementation. '
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAuthGetSessionResponse'
      x-auth-mode: bearer
      x-auth-audience: ADMIN_WEB
      x-contract-status: CONTRACT_SYNCED_IMPLEMENTATION_CANDIDATE
      x-approval-source: 2963e3918bbd62afdfdeb55c60bb2e2221c67ca1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    authAttempt:
      type: apiKey
      in: header
      name: X-Auth-Attempt
      description: Restricted authentication-attempt secret, not a business Bearer or client asserted identity.
    adminBinding:
      type: apiKey
      in: cookie
      name: __Host-pet-admin-attempt
      description: Same-origin Secure HttpOnly SameSite=Strict request binding; checked together with Origin and attempt
        secret, not an extra authentication factor.
  parameters:
    RequestId:
      name: X-Request-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
        minLength: 36
        maxLength: 36
        pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}(?![\s\S])'
      description: 终端完整UUID原值保留；同意图重试复用，不trim或改大小写；内部确定性串不适用本header。
...
  schemas:
...
    AuthErrorEnvelope:
      type: object
      additionalProperties: false
      required:
      - code
      - message
      - data
      - traceId
      properties:
        code:
          type: string
          enum:
          - COMMON_INVALID_ARGUMENT
          - COMMON_UNAUTHORIZED
          - COMMON_FORBIDDEN
          - COMMON_NOT_FOUND
          - COMMON_CONFLICT
          - COMMON_RATE_LIMITED
          - COMMON_INTERNAL_ERROR
          - COMMON_DEPENDENCY_UNAVAILABLE
          - IDEMPOTENCY_KEY_CONFLICT
        message:
          type: string
          minLength: 1
        data:
          type: object
          nullable: true
          enum:
          - null
        traceId:
          type: string
          minLength: 1
...
    AuthMiniAttemptRequest:
      type: object
      additionalProperties: false
      required:
      - purpose
      properties:
        purpose:
          type: string
          enum:
          - WECHAT_LOGIN
          - SMS_LOGIN
          - PASSWORD_LOGIN
          - PASSWORD_RESET
    AuthEmptyRequest:
      type: object
      additionalProperties: false
      properties: {}
...
    CAuthCreateAttemptResponse:
      type: object
      additionalProperties: false
      required:
      - code
      - message
      - data
      - traceId
      properties:
        code:
          type: string
          enum:
          - SUCCESS
        message:
          type: string
          minLength: 1
        data:
          $ref: '#/components/schemas/AuthAttemptCreated'
        traceId:
          type: string
          minLength: 1
...